Speed up task time when nothing needs to be done.
This checks the last modified time of the output file, if it exists, and
compares it to the last modified time of the input file.  If the input
file is newer, run the compilation.

Also, wrap less-compiler with (delay ...), so that we don't instantiate
the compiler until we need it.  This is where the real cost savings
happens.

Finally, add a hook around the compile process.  A project that wants to
use this feature has to add the following to their project.clj:

    :hooks [leiningen.hooks.lesscss]

Overall, the speed gain seems to be about 5x faster on my machine.